move more functions into data_source_meta
