+++ corehq/apps/userreports/reports/builder/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
-import six
+
 DEFAULT_CASE_PROPERTY_DATATYPES = {
     "name": "string",
     "modified_on": "datetime",
@@ -64,27 +64,6 @@ def make_case_property_indicator(property_name, column_id=None, datatype=None):
     }
 
 
-def make_form_question_indicator(question, column_id=None, data_type=None, root_doc=False):
-    """
-    Return a data source indicator configuration (a dict) for the given form
-    question.
-    """
-    path = question['value'].split('/')
-    expression = {
-        "type": "property_path",
-        'property_path': ['form'] + path[2:],
-    }
-    if root_doc:
-        expression = {"type": "root_doc", "expression": expression}
-    return {
-        "type": "expression",
-        "column_id": column_id or question['value'],
-        "display_name": path[-1],
-        "datatype": data_type or get_form_indicator_data_type(question['type']),
-        "expression": expression
-    }
-
-
 def make_multiselect_question_indicator(question, column_id=None):
     path = question['value'].split('/')
     return {
@@ -95,44 +74,6 @@ def make_multiselect_question_indicator(question, column_id=None):
         "select_style": "multiple",
         "choices": [o['value'] for o in question['options']],
     }
-
-
-def make_form_meta_block_indicator(spec, column_id=None, root_doc=False):
-    """
-    Return a data source indicator configuration (a dict) for the given
-    form meta field and data type.
-    """
-    field_name = spec[0]
-    if isinstance(field_name, six.string_types):
-        field_name = [field_name]
-    data_type = spec[1]
-    column_id = column_id or field_name[0]
-    expression = {
-        "type": "property_path",
-        "property_path": ['form', 'meta'] + field_name,
-    }
-    if root_doc:
-        expression = {"type": "root_doc", "expression": expression}
-    return {
-        "type": "expression",
-        "column_id": column_id,
-        "display_name": field_name[0],
-        "datatype": get_form_indicator_data_type(data_type),
-        "expression": expression
-    }
-
-
-def get_form_indicator_data_type(question_type):
-    return {
-        "date": "date",
-        "datetime": "datetime",
-        "Date": "date",
-        "DateTime": "datetime",
-        "Int": "integer",
-        "Double": "decimal",
-        "Text": "string",
-        "string": "string",
-    }.get(question_type, "string")
 
 
 def get_filter_format_from_question_type(question_type):
