+++ corehq/apps/userreports/reports/builder/columns.py
@@ -1,11 +1,11 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 from corehq.apps.userreports.reports.builder import (
-    make_form_meta_block_indicator,
     make_case_property_indicator,
-    make_form_question_indicator,
     make_multiselect_question_indicator,
 )
+from corehq.apps.userreports.app_manager.data_source_meta import make_form_question_indicator, \
+    make_form_meta_block_indicator
 from corehq.apps.userreports.reports.builder.const import (
     UCR_AGG_AVG,
     UCR_AGG_EXPAND,
